2018-12-17  Benjamin De Kosnik  <bdekoz@mozilla.com>
	* src/moz-json-basic.h (extract_histogram_field_median):
	Cast to integer.
	* src/moz-telemetry-x-analyze.cc (radiate_name_by_value):
	Don't print zero value fields.

+++ src/moz-json-basic.h
@@ -252,7 +252,7 @@ extract_histogram_field_median(const rj::Value& v, const string& probe)
 		  auto m2 = vvalues[(vvsize / 2) - 1];
 		  median = (m1 + m2) / 2;
 		}
-	      found = to_string(median);
+	      found = to_string(static_cast<uint>(median));
 	    }
 	}
     }
